get random points fix

+++ cpp/unreal_plugins/SimulationController/Source/SimulationController/CameraAgent.cpp
@@ -194,7 +194,8 @@ std::map<std::string, std::vector<uint8_t>> CameraAgent::getStepInfo() const
     auto step_info_components = Config::get<std::vector<std::string>>("SIMULATION_CONTROLLER.CAMERA_AGENT.STEP_INFO_COMPONENTS");
 
     if (Std::contains(step_info_components, "random_points")) {
-        std::vector<float> component_data = Std::reinterpret_as<float>(action_.at("set_num_random_points"));
+
+        std::vector<uint32> component_data = Std::reinterpret_as<uint32>(action_.at("set_num_random_points"));
         std::vector<float> random_points;
         int num_random_points = static_cast<int>(component_data.at(0));
         for (int i = 0; i < num_random_points; i++) {
